Add en_passant_file column to the games table
Here, "file" is used as the chess-ier synonym for "column".

An en passant capture is only possible on the move immediately following
a two-square advance by an opposing pawn. This column keeps track of
whether or not this applies.

At each move, the appropriate move_to! method will update this value to
the correct file/column number if the move is a two-square advance by
a pawn, and set it to nil/null otherwise.

If an en passant capture is attempted, valid_move? and move_to! (or
their helper methods) will read this value as part of the processes of
verifying the move's legality and capturing the enemy pawn.