Add nth_rank helper method to piece model
As noted in the method comment, this method takes the rank from the
current player's perspective, and returns the corresponding y value.
For example, pawns start on the player's second rank, which is 2 for
White and 7 for Black. Therefore, nth_rank(2) returns 2 if the piece is
white, and 7 if it's black.