Refactor valid_double_move? to use nth_rank
The valid_double_move? method, in the Pawn model, verifies that the pawn
is on its original row, and that neither of the spaces immediately in
front of it is occupied.

The newly-created nth_rank helper method removes the need to explicitly
call the ternary operator to set the values of origin_row and
passed_row.

Because each of these values is used only once, and they can
now be defined non-verbosely, the variables themselves are no longer
particularly useful, and the calls to nth_rank can be put directly in
the methods which use the values.

(Additionally, the method's last two lines are refactored into one.)