Update valid_capture? method for en passant
Previously, the Pawn model's valid_capture method simply returned true
if the destination square contained an enemy piece, and false otherwise.

With this update, the method continues to return true if a standard
capture is possible, but also returns true if an enemy pawn can be
captured en passant. This will be the case if:

- the capturing pawn starts on its fifth rank, and
- the game's immediately-preceding move was a pawn advancing two squares
  along the destination file (column).

If the latter applies, the game's newly-created en_passant_file
attribute will equal the x coordinate of the destination square.