Update en_passant_file in Piece's move_to! method
If the current move is not being made by a pawn, en passant capture is
not possible on the next move. Set the game's en_passant_file attribute
to nil to reflect this. (For moves by pawns, the move_to! method in the
Pawn class sets this value appropriately.)

This commit also includes the following minor tweaks:
- Reload the game before retrieving en_passant_file in the pawn's
  valid_capture? method.
- Remove a commented-out line of debugging code.