Add basic test coverage
Set up a game, make a series of moves leading to an en passant capture,
and verify that the capture is valid and the captured piece is
destroyed.

Then, make a series of moves that would lead to the possibility of en
passant capture, but make another move next instead. Verify that the
en passant capture is invalid, and that the no-longer-vulnerable pawn
is still present.