Add non-pawn and no-two-square-advance tests
Verify that en passant capture cannot be made by a non-pawn, by
attempting to take a pawn that just advanced two squares by moving
a bishop to the square it passed through. This is a legal bishop move,
but leaves the pawn uncaptured.

Verify that en passant capture cannot be made on a pawn that has made
two separate one-square advances, rather than a single two-square
advance.