Add column, current_player, to the games table
Declare an enum attribute in the Game model, utilizing the
newly created current_player. Since white always goes first,
it is the first item in the array. The values in the array map to
integers in the database. This corresponds with the default value
that was placed for current_player in the db.