Enhancements for displaying rank and file
* Adjust master.js, create a new function gameStartMessage, so as to
indicate whose turn it is at the beginning of a game.
* Adjust the games#show page and add bootstrap grid classes to the
board-booyah-box. Add a script to call to display the gameStartMessage.
* Adjust game.css.scss and add further optimizations for mobile devices
* Add CSS classes for rank and file indicators.
* Adjust _chess_board.html.erb, add table columns to the chessboard,
and use the left-most column to indicate rank. Collapse the borders for
those new cells. Add lower case letters to indicate file columns.
* Add the add_squash_class method inside game_helper.rb to add a class
for the cells to be collapsed.
* Adjust the CSS for the checkerboard pattern so the board still looks
correct.