Update gems
Unlike the previous commit, this does not fix any security
vulnerabilities that are known at this time.

I tested that logging in as two different users and playing a few moves
still works as expected, but didn't go any further than that.

Feel free to revert if any issues are discovered.